[BRANCH] The BRANCH argument is introduced
The BRANCH argument has now been introduced so we
can also manage the associated application branches,
just as we did with the FLOW arguments:

1) The BRANCH_INDEX constant can be used to fetch
   the BRANCH_FLAG from the program arguments provided
   by the user.

2) If the BRANCH_INDEX contains something else we throw
   an InvalidUseException to make sure that the application
   terminates accordingly and to alert the user. The 1 index
   is now locked to the BRANCH argument.

3) The BRANCH task will now be logged, under the PROCESSING_TASK
   constant, which will be extended under the upcoming commits
   with more information.

This commit also includes implementation and testing for:

1) The Extraction feature, has been updated to hold the
   appropriate branch tasks, includes a new function which is
   used to hide away the number of upcoming 'when' statements
   that would otherwise clutter the code.

Finally, there were some updates made to the previous test cases,
as the program argument list required an update, but latter arguments
are ignored as the negative test cases fail before they are even
validated. Others are updated to include the associated branch.