[RESOURCE] The RESOURCE argument is introduced
The RESOURCE argument has now been introduced so we
can also manage the associated application resource,
just as we did with the BRANCH arguments:

1) The RESOURCE_INDEX constant can be used to fetch
   the RESOURCE_FLAG from the program arguments provided
   by the user.

2) If the RESOURCE_INDEX contains something else we throw
   an InvalidUseException to make sure that the application
   terminates accordingly and to alert the user. The 2 index
   is now locked to the RESOURCE argument.

3) The RESOURCE task will now be logged, under the PROCESSING_TASK
   constant, which will be extended under the upcoming commits
   with more information.

This commit also includes implementation and testing for:

1) The Extraction feature, has been updated to hold the
   appropriate resource tasks, includes a new function which is
   used to hide away the number of upcoming 'when' statements
   that would otherwise clutter the code.

Finally, there were some updates made to the previous test cases,
as the program argument list required an update, but latter arguments
are ignored as the negative test cases fail before they are even
validated. Others are updated to include the associated branch.