[PATH] The PATH argument is introduced
The PATH argument has now been introduced so we can
use a target directory for the application to process
files, just as we did with previous arguments:

1) The PATH_INDEX constant can be used to fetch the
   actual path from the program arguments provided by
   the user.

2) If the PATH_INDEX contains something else we throw
   an DirectoryUnavailableException to make sure that
   the application terminates accordingly and to alert
   the user. The '3' index is now locked to the PATH
   argument.

3) The PATH itself will now be logged, under the
   PROCESSING_TASK constant, which will be extended
   under the upcoming commits with more information.

This commit also includes implementation and testing for:

1) The FileProcessing feature, which has been added to
   hold the upcoming functions for I/O management, includes
   a new function which is used to validate a directory
   path provided.

Finally, there were some updates made to the previous test
cases, as the program argument list required an update, but
latter arguments are ignored as the negative test cases fail
before they are even validated. Others are updated to include
the directory path.