[EXTENSION] The EXTENSION arguments are introduced
The EXTENSION arguments have now been introduced,
so the application can target specific files to
handle, just as we did with the previous arguments:

1) The FROM_INDEX & TO_INDEX constants can be used
   to fetch the specific EXTENSION that the
   application should focus on, from the program
   arguments provided by the user. The respective
   will have different purposes and will be less
   important depending on FLOW and BRANCH.

2) If the EXTENSION are out of bounds, depending on
   the MEDIA_SUPPORT collection, we throw an
   InvalidUseException to make sure that the application
   terminates accordingly and to alert the user. The
   '4' & '5' indexes are now locked to the EXTENSION
   arguments.

3) The EXTENSIONS will now be logged, under the
   PROCESSING_TASK constant, which will be extended
   under the upcoming commits with more information.

Finally, there were some updates made to the previous test
cases, as the program argument list required an update, but
latter arguments are ignored as the negative test cases fail
before they are even validated. Others are updated to include
the extensions as they are currently used.