[AXIS] The AXIS arguments are introduced
The AXIS arguments have now been introduced, so
the application will be able to crop an image
starting from a specific Y and X pixel point,
just as we did with the previous arguments:

1) The LEFT_Y_AXIS & LEFT_X_AXIS constants are
   used to crop an image, starting from these
   pixel points, from the program arguments
   provided by the user.

2) If the AXES are out of bounds, we throw an
   InvalidUseException to make sure that the
   application terminates accordingly and to
   alert the user. The '6' & '7' indexes are
   now locked to the AXIS arguments.

3) The AXES will now be logged, under the
   PROCESSING_ATTRIBUTES constant, which will
   be extended under the upcoming commits with
   more information.

Finally, there were some updates made to the previous
test cases, as the program argument list required an
update, but latter arguments are ignored as the negative
test cases fail before they are even validated. Others
are updated to include the axes as they are currently used.