[TIMESTAMP] The TIMESTAMP argument is introduced
The TIMESTAMP argument has now been introduced so we
can target a specific creation, or modification date,
just as we did with the previous arguments:

1) The TIMESTAMP constant can be used to fetch the
   TIMESTAMP_FLAG from the program arguments provided
   by the user.

2) If the TIMESTAMP_INDEX contains something else we throw
   an InvalidUseException to make sure that the application
   terminates accordingly and to alert the user. The '8'
   index is now locked to the TIMESTAMP argument.

3) The TIMESTAMP task will now be logged, under the
   PROCESSING_TASK constant, which will be extended under
   the upcoming commits with more information.

Finally, there were some updates made to the previous test
cases, as the program argument list required an update, but
latter arguments are ignored as the negative test cases fail
before they are even validated. Others are updated to include
the associated timestamp.