[BOUNDARY] The BOUNDARY argument is introduced
The BOUNDARY argument has now been introduced so we
can determine at what percentage a process should
terminate at, just as we did with the previous arguments:

1) The BOUNDARY_INDEX constant can be used to fetch
   the BOUNDARY value from the program arguments provided
   by the user.

2) If the BOUNDARY_INDEX contains something else we throw
   an InvalidUseException to make sure that the application
   terminates accordingly and to alert the user. The '9'
   index is now locked to the BOUNDARY argument.

3) The BOUNDARY task will now be logged, under the
   PROCESSING_TASK constant, and this will be the final
   update to this constant.

Finally, there were some updates made to the previous test
cases, as the program argument list required an update, but
latter arguments are ignored as the negative test cases fail
before they are even validated. Others are updated to include
the associated boundary.